+++ src/r_voxel.c
@@ -1381,9 +1381,6 @@ boolean VX_ProjectWeaponVoxel(const pspdef_t *const psp,
   thing.sprite = psp->state->sprite;
   thing.frame  = psp->state->frame;
 
-  if (translucent)
-  { thing.flags |= MF_TRANSLUCENT; }
-
   if (POWER_RUNOUT(viewplayer->powers[pw_invisibility]) && !beta_emulation)
   { thing.flags |= MF_SHADOW; }