+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 
 - **Improved _Hitbox Thing Lighting_**
 - **Reduced sensitivity of _Vertical Target Lock-on_**
+- **Reduced opacity of hitscan trails**
 - **Made `TURBO` cheat display a help message**
 
 ## Bug Fixes
 
 - **Crash when changing _[HUD] Layout_ without loading into a map since launching the game**
+- **Translucent weapon flashes and hitscan trails not being translucent if _Sprite Translucency_ were disabled**
 - **Keyframe time becoming misaligned when rewinding**
 - **_Hitbox Thing Lighting_ not applying to weapon voxels**
 - **_Vertical Target Lock-on_ menu item not being disabled when _Free Look_ were enabled**